update dotnet 7 to 8 add mappers add chenge code in files (app components and jwt.interceptors for date expire token)

+++ client/src/app/services/account.service.ts
@@ -1,11 +1,13 @@
 import { HttpClient } from '@angular/common/http';
-import { Injectable, inject } from '@angular/core';
+import { Injectable, PLATFORM_ID, inject } from '@angular/core';
 import { RegisterUser } from '../model/register-user.model';
 import { BehaviorSubject, Observable, map } from 'rxjs';
 import { User } from '../model/user.model';
 import { LoginUser } from '../model/login.user.model';
 import { Router } from '@angular/router';
 import { environment } from '../../environments/environment.development';
+import { Token } from '@angular/compiler';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
@@ -16,6 +18,7 @@ export class AccountService {
 
   http = inject(HttpClient);
   router = inject(Router);
+  platformId = inject(PLATFORM_ID);
 
   // private readonly baseApiUrl: string = 'http://localhost:5000/api/account/';
   private readonly baseApiUrl = environment.apiUrl + 'account/';
@@ -50,22 +53,34 @@ export class AccountService {
       })
     );
   }
+  // TODO add method getUser To get user from api
+  getUser(): Observable<User | null> {
+    return this.http.get<User>(environment.apiUrl + 'user/get-by-id').pipe(
+      map((userResponce: User | null) => {
+
+        if (userResponce)
+          return userResponce
+
+        return null;
+      })
+    );
+  }
 
   setCurrentUser(user: User): void {
     this.currentUserSourse.next(user);
 
-    localStorage.setItem('user', JSON.stringify(user));
+    // localStorage.setItem('user', JSON.stringify(user)); // before date expire token (video 21)
+    localStorage.setItem('token', user.token);
 
     this.router.navigateByUrl('');
   }
 
   LogoutUser(): void {
     this.currentUserSourse.next(null);
-
-    localStorage.clear();
-
-    // this.router.navigateByUrl('account/login');
+    if (isPlatformBrowser(this.platformId))
+      localStorage.clear(); // remove token in localStorage
+    // localStorage.removeItem('user'); // before date expire token (video 21)
+    localStorage.removeItem('token');
     this.router.navigateByUrl('account/login');
-
   }
 }
